:construction: (RUMM-395): Merge modules with main library
