+++ LICENSE-3rdparty.csv
@@ -1,11 +1,37 @@
 Component,Origin,License,Copyright
+import,androidx.activity,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.annotation,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.appcompat,Apache-2.0,Copyright 2018 The Android Open Source Project
 import,androidx.arch.core,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.asynclayoutinflater,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.cardview,Apache-2.0,Copyright 2018 The Android Open Source Project
 import,androidx.collection,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.coordinatorlayout,Apache-2.0,Copyright 2018 The Android Open Source Project
 import,androidx.core,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.cursoradapter,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.customview,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.documentfile,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.drawerlayout,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.fragment,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.interpolator,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.legacy,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.lifecycle,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.loader,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.localbroadcastmanager,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.navigation,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.print,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.recyclerview,Apache-2.0,Copyright 2018 The Android Open Source Project
 import,androidx.room,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.savedstate,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.slidingpanelayout,Apache-2.0,Copyright 2018 The Android Open Source Project
 import,androidx.sqlite,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.swiperefreshlayout,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.transition,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.vectordrawable,Apache-2.0,Copyright 2018 The Android Open Source Project
 import,androidx.versionedparcelable,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,androidx.viewpager,Apache-2.0,Copyright 2018 The Android Open Source Project
 import,androidx.work,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,com.google.android.material,Apache-2.0,__
 import,com.google.code.gson,Apache-2.0,Copyright 2008 Google Inc
 import,com.google.guava,Apache-2.0,Copyright 2009 The Guava Authors
 import,com.jakewharton.timber,Apache-2.0,Copyright 2013 Jake Wharton
@@ -18,18 +44,9 @@ import,io.opentracing,Apache-2.0,Copyright 2016-2017 The OpenTracing Authors
 import,io.opentracing.contrib,Apache-2.0,Copyright 2016-2017 The OpenTracing Authors
 import,org.jetbrains,Apache-2.0,Copyright 2010-2019 JetBrains s.r.o. and Kotlin Programming Language contributors
 import,org.jetbrains.kotlin,Apache-2.0,Copyright 2010-2019 JetBrains s.r.o. and Kotlin Programming Language contributors
-import,org.slf4j,MIT,Copyright (c) 2004-2017 QOS.ch
-import,org.msgpack,Boost Software License-1.0,Copyright 2019 the original author or authors
-import,org.ow2,BSD 3-Clause Clear License,2000-2011 INRIA, France Telecom
-import,org.sonatype.oss,Apache-2.0,Copyright (c) 2007-2020 Sonatype, Inc. All rights reserved.
-import(test),androidx.annotation,Apache-2.0,Copyright 2019 The Android Open Source Project
-import(test),androidx.benchmark,Apache-2.0,Copyright 2019 The Android Open Source Project
-import(test),androidx.lifecycle,Apache-2.0, Version 2.0,Copyright 2019 The Android Open Source Project
-import(test),androidx.test,Apache-2.0,Version 2.0,Copyright 2019 The Android Open Source Project
-import(test),androidx.test.ext,Apache-2.0,Copyright 2018 The Android Open Source Project
+import,org.jetbrains.kotlinx,Apache-2.0,Copyright 2010-2019 JetBrains s.r.o. and Kotlin Programming Language contributors
 import(test),com.github.xgouchet.Elmyr,MIT,Copyright 2017-2019 Xavier F. Gouchet
 import(test),com.nhaarman.mockitokotlin2,MIT,"Copyright (c) 2018 Niek Haarman, Copyright (c) 2007 Mockito contributors"
-import(test),com.google.auto,Apache-2.0,"Copyright 2013 Google, Inc."
 import(test),junit,EPL-1.0,Copyright © 2002-2019 JUnit
 import(test),net.bytebuddy,Apache-2.0,Copyright 2014 - 2019 Rafael Winterhalter
 import(test),net.sf.kxml,MIT,"Copyright (c) 2002,2003, Stefan Haustein, Oberhausen, Rhld., Germany"
@@ -45,8 +62,8 @@ import(test),org.mockito,MIT,Copyright (c) 2007 Mockito contributors
 import(test),org.objenesis,Apache-2.0,"Copyright (c) 2003-2013, Objenesis Team and all contributors"
 import(test),org.opentest4j,Apache-2.0,Copyright 2015-2018 the original author or authors
 import(test),org.robolectric,Apache-2.0,Copyright 2015-2018 the original author or authors
-build,io.gitlab.arturbosch.detekt,Apache-2.0,Copyright 2016-2019 the original author or authors
 build,com.android.tools.build,Apache-2.0,Copyright (C) 2013 The Android Open Source Project
 build,com.android.tools.lint,Apache-2.0,Copyright (C) 2013 The Android Open Source Project
 build,com.pinterest.ktlint,MIT,"Copyright 2019 Pinterest Inc, Copyright 2016-2019 Stanley Shyiko"
+build,io.gitlab.arturbosch.detekt,Apache-2.0,Copyright 2016-2019 the original author or authors
 build,org.jetbrains.dokka,Apache-2.0,"Copyright 2014-2019 JetBrains s.r.o. and Dokka project contributors."
