--- dd-sdk-android-androidx/apiSurface
@@ -1,6 +1,0 @@
-class com.datadog.android.androidx.fragment.FragmentViewTrackingStrategy : com.datadog.android.rum.tracking.ActivityLifecycleTrackingStrategy, com.datadog.android.rum.tracking.ViewTrackingStrategy
-  constructor(Boolean)
-  override fun onActivityResumed(android.app.Activity)
-  override fun onActivityPaused(android.app.Activity)
-class com.datadog.android.androidx.tracking.JetpackViewAttributesProvider : com.datadog.android.rum.tracking.ViewAttributesProvider
-  override fun extractAttributes(android.view.View, MutableMap<String, Any?>)
