--- dd-sdk-android-androidx/build.gradle.kts
@@ -1,103 +1,0 @@
-/*
- * Unless explicitly stated otherwise all pomFilesList in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-import com.datadog.gradle.Dependencies
-import com.datadog.gradle.config.AndroidConfig
-import com.datadog.gradle.config.bintrayConfig
-import com.datadog.gradle.config.dependencyUpdateConfig
-import com.datadog.gradle.config.detektConfig
-import com.datadog.gradle.config.jacocoConfig
-import com.datadog.gradle.config.javadocConfig
-import com.datadog.gradle.config.junitConfig
-import com.datadog.gradle.config.kotlinConfig
-import com.datadog.gradle.config.ktLintConfig
-import com.datadog.gradle.config.publishingConfig
-import com.datadog.gradle.testImplementation
-
-plugins {
-    id("com.android.library")
-    kotlin("android")
-    kotlin("android.extensions")
-    `maven-publish`
-    id("com.github.ben-manes.versions")
-    id("io.gitlab.arturbosch.detekt")
-    id("org.jlleitschuh.gradle.ktlint")
-    id("thirdPartyLicences")
-    id("apiSurface")
-    id("org.jetbrains.dokka")
-    id("com.jfrog.bintray")
-    id("de.mobilej.unmock")
-    jacoco
-}
-
-android {
-    compileSdkVersion(AndroidConfig.TARGET_SDK)
-    buildToolsVersion(AndroidConfig.BUILD_TOOLS_VERSION)
-
-    defaultConfig {
-        minSdkVersion(AndroidConfig.MIN_SDK)
-        targetSdkVersion(AndroidConfig.TARGET_SDK)
-        versionCode = AndroidConfig.VERSION.code
-        versionName = AndroidConfig.VERSION.name
-    }
-
-    sourceSets.named("main") {
-        java.srcDir("src/main/kotlin")
-    }
-    sourceSets.named("test") {
-        java.srcDir("src/test/kotlin")
-    }
-    sourceSets.named("androidTest") {
-        java.srcDir("src/androidTest/kotlin")
-    }
-
-    testOptions {
-        unitTests.isReturnDefaultValues = true
-    }
-
-    lintOptions {
-        isWarningsAsErrors = true
-        isAbortOnError = true
-        isCheckReleaseBuilds = false
-        isCheckGeneratedSources = true
-    }
-}
-
-dependencies {
-    api(project(":dd-sdk-android"))
-    implementation(Dependencies.Libraries.Kotlin)
-    implementation(Dependencies.Libraries.AndroidXAppCompat)
-    implementation(Dependencies.Libraries.AndroidXRecyclerView)
-
-    testImplementation(project(":tools:unit"))
-    testImplementation(Dependencies.Libraries.JUnit5)
-    testImplementation(Dependencies.Libraries.TestTools)
-    testImplementation(Dependencies.Libraries.OkHttpMock)
-
-    unmock(Dependencies.Libraries.Robolectric)
-
-    detekt(project(":tools:detekt"))
-    detekt(Dependencies.Libraries.DetektCli)
-}
-
-unMock {
-    keep("android.os.BaseBundle")
-    keep("android.os.Bundle")
-    keep("android.os.Parcel")
-    keepStartingWith("com.android.internal.util.")
-    keepStartingWith("android.util.")
-    keep("android.content.ComponentName")
-}
-
-kotlinConfig()
-detektConfig()
-ktLintConfig()
-junitConfig()
-jacocoConfig()
-javadocConfig()
-dependencyUpdateConfig()
-publishingConfig("${rootDir.canonicalPath}/repo")
-bintrayConfig()
