+++ dd-sdk-android/build.gradle.kts
@@ -18,6 +18,7 @@ import com.datadog.gradle.config.junitConfig
 import com.datadog.gradle.config.kotlinConfig
 import com.datadog.gradle.config.ktLintConfig
 import com.datadog.gradle.config.publishingConfig
+import com.datadog.gradle.implementation
 import com.datadog.gradle.testImplementation
 
 plugins {
@@ -107,8 +108,13 @@ dependencies {
     implementation(Dependencies.Libraries.Gson)
     implementation(Dependencies.Libraries.Kotlin)
     implementation(Dependencies.Libraries.OkHttp)
+
     implementation(Dependencies.Libraries.AndroidXWorkManager)
     implementation(Dependencies.Libraries.AndroidXCore)
+    implementation(Dependencies.Libraries.AndroidXAppCompat)
+    implementation(Dependencies.Libraries.AndroidXRecyclerView)
+    implementation(Dependencies.Libraries.AndroidXNavigation)
+
     api(Dependencies.Libraries.TracingOt)
 
     testImplementation(project(":tools:unit"))
