+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/utils/ViewUtils.kt
@@ -1,4 +1,4 @@
-package com.datadog.android.androidx.fragment.internal
+package com.datadog.android.core.internal.utils
 
 internal fun Any.resolveViewName(): String {
     return javaClass.canonicalName ?: javaClass.simpleName
