+++ dd-sdk-android/src/main/kotlin/com/datadog/android/rum/internal/tracking/FragmentLifecycleCallbacks.kt
@@ -1,0 +1,20 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+package com.datadog.android.rum.internal.tracking
+
+import android.app.Activity
+
+internal interface FragmentLifecycleCallbacks<T : Activity> {
+
+    // region Lifecycle
+
+    fun register(activity: T)
+
+    fun unregister(activity: T)
+
+    // end region
+}
