+++ dd-sdk-android/src/main/kotlin/com/datadog/android/rum/tracking/JetpackViewAttributesProvider.kt
@@ -1,10 +1,9 @@
-package com.datadog.android.androidx.tracking
+package com.datadog.android.rum.tracking
 
 import android.content.res.Resources
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.datadog.android.rum.RumAttributes
-import com.datadog.android.rum.tracking.ViewAttributesProvider
 
 /**
  * Provides extra attributes for the touch target View.
