+++ sample/java/src/main/java/com/datadog/android/sample/SampleApplication.java
@@ -17,7 +17,7 @@ import androidx.multidex.MultiDex;
 
 import com.datadog.android.Datadog;
 import com.datadog.android.DatadogConfig;
-import com.datadog.android.androidx.fragment.FragmentViewTrackingStrategy;
+import com.datadog.android.rum.tracking.FragmentViewTrackingStrategy;
 import com.datadog.android.log.Logger;
 import com.datadog.android.rum.GlobalRum;
 import com.datadog.android.rum.RumMonitor;
