+++ settings.gradle.kts
@@ -37,8 +37,6 @@ pluginManagement {
 
 include(":dd-sdk-android")
 include(":dd-sdk-android-timber")
-include(":dd-sdk-android-androidx")
-include(":dd-sdk-android-navigation")
 
 include(":instrumented:benchmark")
 include(":instrumented:integration")
